decouple and organize imports and exports
