+++ plugin.ts
@@ -2,54 +2,33 @@ import type {
     Context,
     MiddlewareFn,
 } from "https://lib.deno.dev/x/grammy@1.x/mod.ts";
-import type {
-    LocaleNegotiator,
+import { negotiateLanguages } from "npm:@fluent/langneg@0.7.0";
+import { createDebug } from "jsr:@grammyjs/debug@0.2.1";
+import { isValidLocale } from "./utilities.ts";
+import {
+    FormatAdapter,
     Locales,
     LocalesTypings,
     MessageKey,
     Messages,
-    NegotiatorResult,
-    TranslateFunction,
 } from "./types.ts";
-import { isValidLocale } from "./utilities.ts";
-import { negotiateLanguages } from "npm:@fluent/langneg@0.7.0";
-import { createDebug } from "jsr:@grammyjs/debug@0.2.1";
 
 const debug = createDebug("grammy:i18n");
 
-/**
- * A format adapter is an abstraction that provides translate capabilities to
- * any localization format. Format adapters helps enable localization regardless
- * of the localization format used. Format adapters should manage the
- * translation resources and expose a translate function that can be called from
- * the i18n instance.
- */
-export interface FormatAdapter<
-    LT extends LocalesTypings,
-> {
-    /**
-     * Get the list of locales registered in the adapter.
-     */
-    locales: string[];
-    /**
-     * Formats and returns a message string if the message exists.
-     *
-     * @param locale Locale to use when translating.
-     * @param messageKey Message key to be used.
-     * @param args Variables to be passed for formatting the message data.
-     */
-    translate<
-        L extends Locales<LT>,
-        MK extends MessageKey<LT, Messages<LT>>,
-    >(
-        locale: L,
-        messageKey: MK,
-        ...args: Messages<LT>[MK] extends never ? []
-            : Messages<LocalesTypings>[string] extends Messages<LT>[MK]
-                ? [variables?: Messages<LT>[MK]]
-            : [variables: Messages<LT>[MK]]
-    ): string | undefined;
-}
+export type NegotiatorResult = string | undefined;
+export type LocaleNegotiator<C extends Context> = (
+    ctx: C,
+) => NegotiatorResult | Promise<NegotiatorResult>;
+
+export type TranslateFunction<LT extends LocalesTypings> = <
+    MK extends MessageKey<LT, Messages<LT>>,
+>(
+    messageKey: MK,
+    ...args: Messages<LT>[MK] extends never ? []
+        : Messages<LocalesTypings>[string] extends Messages<LT>[MK]
+            ? [variables?: Messages<LT>[MK]]
+        : [variables: Messages<LT>[MK]]
+) => string;
 
 /**
  * Context flavor for the outside middleware tree. Installs `ctx.translate` and
