+++ types.ts
@@ -1,14 +1,4 @@
-import type { Context } from "https://lib.deno.dev/x/grammy@1.x/mod.ts";
-
-export type KeyOf<T> = string & keyof T;
-export type StringWithSuggestions<S extends string> =
-    | string & Record<never, never>
-    | S;
-
-export type NegotiatorResult = string | undefined;
-export type LocaleNegotiator<C extends Context> = (
-    ctx: C,
-) => NegotiatorResult | Promise<NegotiatorResult>;
+type KeyOf<T> = string & keyof T;
 
 export type LocalesTypings<
     L extends string = string,
@@ -34,15 +24,39 @@ export type MessageKey<
     M extends Messages<LT>,
 > = KeyOf<M>;
 
-export type TranslateFunction<LT extends LocalesTypings> = <
-    MK extends MessageKey<LT, Messages<LT>>,
->(
-    messageKey: MK,
-    ...args: Messages<LT>[MK] extends never ? []
-        : Messages<LocalesTypings>[string] extends Messages<LT>[MK]
-            ? [variables?: Messages<LT>[MK]]
-        : [variables: Messages<LT>[MK]]
-) => string;
+/**
+ * A format adapter is an abstraction that provides translate capabilities to
+ * any localization format. Format adapters helps enable localization regardless
+ * of the localization format used. Format adapters should manage the
+ * translation resources and expose a translate function that can be called from
+ * the i18n instance.
+ */
+export interface FormatAdapter<
+    LT extends LocalesTypings,
+> {
+    /**
+     * Get the list of locales registered in the adapter.
+     */
+    locales: string[];
+    /**
+     * Formats and returns a message string if the message exists.
+     *
+     * @param locale Locale to use when translating.
+     * @param messageKey Message key to be used.
+     * @param args Variables to be passed for formatting the message data.
+     */
+    translate<
+        L extends Locales<LT>,
+        MK extends MessageKey<LT, Messages<LT>>,
+    >(
+        locale: L,
+        messageKey: MK,
+        ...args: Messages<LT>[MK] extends never ? []
+            : Messages<LocalesTypings>[string] extends Messages<LT>[MK]
+                ? [variables?: Messages<LT>[MK]]
+            : [variables: Messages<LT>[MK]]
+    ): string | undefined;
+}
 
 export interface ResourceLoadable<T> {
     /**
